+++ pkg/apiserver/user/sqlauth/sqlauth.go
@@ -1,10 +1,6 @@
 package sqlauth
 
 import (
-	"encoding/json"
-	"regexp"
-	"strings"
-
 	"github.com/joomcode/errorx"
 	"go.uber.org/fx"
 
@@ -48,30 +44,6 @@ func (a *Authenticator) Authenticate(f user.AuthenticateForm) (*utils.SessionUse
 		// tidb.ErrTiDBAuthFailed
 		return nil, err
 	}
-	defer utils.CloseTiDBConnection(db) //nolint:errcheck
-
-	// Check dashboard privileges
-	// 1. Check whether TiDB SEM is enabled
-	resData, err := a.tidbClient.SendGetRequest("/config")
-	if err != nil {
-		return nil, user.ErrSignInOther.WrapWithNoMessage(err)
-	}
-	var config tidbSecurityConfig
-	err = json.Unmarshal(resData, &config)
-	if err != nil {
-		return nil, user.ErrSignInOther.WrapWithNoMessage(err)
-	}
-	// 2. Get grants
-	var grantRows []string
-	err = db.Raw("show grants for current_user()").Find(&grantRows).Error
-	if err != nil {
-		return nil, user.ErrSignInOther.WrapWithNoMessage(err)
-	}
-	grants := parseCurUserGrants(grantRows)
-	// 3. Check
-	if !checkDashboardPrivileges(grants, config.Security.EnableSem) {
-		return nil, user.ErrLackPrivileges.NewWithNoMessage()
-	}
 
 	return &utils.SessionUser{
 		Version:      utils.SessionVersion,
@@ -82,84 +54,4 @@ func (a *Authenticator) Authenticate(f user.AuthenticateForm) (*utils.SessionUse
 		IsShareable:  true,
 		IsWriteable:  true,
 	}, nil
-}
-
-// Currently, There are 2 kinds of grant output format in TiDB:
-// - GRANT [grants] ON [db.table] TO [user]
-// - GRANT [roles] TO [user]
-// Examples:
-// - GRANT PROCESS,SHOW DATABASES,CONFIG ON *.* TO 'dashboardAdmin'@'%'
-// - GRANT SYSTEM_VARIABLES_ADMIN,RESTRICTED_VARIABLES_ADMIN,RESTRICTED_STATUS_ADMIN,RESTRICTED_TABLES_ADMIN ON *.* TO 'dashboardAdmin'@'%'
-// - GRANT ALL PRIVILEGES ON *.* TO 'dashboardAdmin'@'%'
-// - GRANT `app_read`@`%` TO `test`@`%`
-func parseCurUserGrants(grantRows []string) []string {
-	grants := make([]string, 0)
-	grantRegex := regexp.MustCompile(`GRANT (.+) ON`)
-
-	for _, row := range grantRows {
-		m := grantRegex.FindStringSubmatch(row)
-		if len(m) == 2 {
-			curRowGrants := strings.Split(m[1], ",")
-			grants = append(grants, curRowGrants...)
-		}
-	}
-
-	return grants
-}
-
-// Following base privileges are required
-// - ALL PRIVILEGES
-// - or
-// - PROCESS
-// - SHOW DATABASES
-// - CONFIG
-// - SYSTEM_VARIABLES_ADMIN or SUPER (SUPER includes SYSTEM_VARIABLES_ADMIN)
-// - DASHBOARD_CLIENT or SUPER (SUPER includes DASHBOARD_CLIENT)
-// When TiDB SEM is enabled, following extra privileges are required
-// - RESTRICTED_VARIABLES_ADMIN
-// - RESTRICTED_TABLES_ADMIN
-// - RESTRICTED_STATUS_ADMIN
-func checkDashboardPrivileges(grants []string, enableSEM bool) bool {
-	hasPriv := func(priv string) bool {
-		for _, grant := range grants {
-			if priv == grant {
-				return true
-			}
-		}
-		return false
-	}
-
-	hasAllPriv := hasPriv("ALL PRIVILEGES")
-	if !hasAllPriv {
-		if !hasPriv("PROCESS") {
-			return false
-		}
-		if !hasPriv("SHOW DATABASES") {
-			return false
-		}
-		if !hasPriv("CONFIG") {
-			return false
-		}
-		hasSuperPriv := hasPriv("SUPER")
-		if !hasSuperPriv {
-			if !hasPriv("SYSTEM_VARIABLES_ADMIN") {
-				return false
-			}
-			if !hasPriv("DASHBOARD_CLIENT") {
-				return false
-			}
-		}
-	}
-	if enableSEM {
-		if !hasPriv("RESTRICTED_VARIABLES_ADMIN") {
-			return false
-		}
-		if !hasPriv("RESTRICTED_TABLES_ADMIN") {
-			return false
-		}
-		if !hasPriv("RESTRICTED_TABLES_ADMIN") {
-			return false
-		}
-	}
-	return true
 }
