+++ pkg/apiserver/utils/verify_sql_user.go
@@ -19,7 +19,6 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/pingcap/tidb-dashboard/pkg/apiserver/utils"
 	"github.com/pingcap/tidb-dashboard/pkg/tidb"
 )
 
@@ -43,7 +42,7 @@ func VerifySQLUser(tidbClient *tidb.Client, user, password string) error {
 	if err != nil {
 		return err
 	}
-	defer utils.CloseTiDBConnection(db) //nolint:errcheck
+	defer CloseTiDBConnection(db) //nolint:errcheck
 
 	// Check privileges
 	// 1. Check whether TiDB SEM is enabled
@@ -64,58 +63,7 @@ func VerifySQLUser(tidbClient *tidb.Client, user, password string) error {
 	}
 	grants := parseCurUserGrants(grantRows)
 	// 3. Check
-	// Following base privileges are required
-	// - ALL PRIVILEGES
-	// - or
-	// - PROCESS
-	// - SHOW DATABASES
-	// - CONFIG
-	// - SYSTEM_VARIABLES_ADMIN or SUPER (SUPER includes SYSTEM_VARIABLES_ADMIN)
-	// - DASHBOARD_CLIENT or SUPER (SUPER includes DASHBOARD_CLIENT)
-	// When TiDB SEM is enabled, following extra privileges are required
-	// - RESTRICTED_VARIABLES_ADMIN
-	// - RESTRICTED_TABLES_ADMIN
-	// - RESTRICTED_STATUS_ADMIN
-	hasPriv := func(priv string) bool {
-		for _, grant := range grants {
-			if priv == grant {
-				return true
-			}
-		}
-		return false
-	}
-	fullfillPrivileges := false
-	if hasPriv("ALL PRIVILEGES") {
-		fullfillPrivileges = true
-	} else {
-		if !hasPriv("PROCESS") {
-			fullfillPrivileges = false
-		} else if !hasPriv("SHOW DATABASES") {
-			fullfillPrivileges = false
-		} else if !hasPriv("CONFIG") {
-			fullfillPrivileges = false
-		} else {
-			if hasPriv("SUPER") {
-				fullfillPrivileges = true
-			} else {
-				if !hasPriv("SYSTEM_VARIABLES_ADMIN") {
-					fullfillPrivileges = false
-				} else if !hasPriv("DASHBOARD_CLIENT") {
-					fullfillPrivileges = false
-				} else {
-					fullfillPrivileges = true
-				}
-			}
-		}
-	}
-	if config.Security.EnableSem {
-		fullfillPrivileges =
-			fullfillPrivileges &&
-				hasPriv("RESTRICTED_VARIABLES_ADMIN") &&
-				hasPriv("RESTRICTED_TABLES_ADMIN") &&
-				hasPriv("RESTRICTED_STATUS_ADMIN")
-	}
-	if !fullfillPrivileges {
+	if !checkDashboardPrivileges(grants, config.Security.EnableSem) {
 		return errors.New("miss required privileges")
 	}
 	return nil
@@ -142,4 +90,61 @@ func parseCurUserGrants(grantRows []string) []string {
 	}
 
 	return grants
+}
+
+// Following base privileges are required
+// - ALL PRIVILEGES
+// - or
+// - PROCESS
+// - SHOW DATABASES
+// - CONFIG
+// - SYSTEM_VARIABLES_ADMIN or SUPER (SUPER includes SYSTEM_VARIABLES_ADMIN)
+// - DASHBOARD_CLIENT or SUPER (SUPER includes DASHBOARD_CLIENT)
+// When TiDB SEM is enabled, following extra privileges are required
+// - RESTRICTED_VARIABLES_ADMIN
+// - RESTRICTED_TABLES_ADMIN
+// - RESTRICTED_STATUS_ADMIN
+func checkDashboardPrivileges(grants []string, enableSEM bool) bool {
+	hasPriv := func(priv string) bool {
+		for _, grant := range grants {
+			if priv == grant {
+				return true
+			}
+		}
+		return false
+	}
+
+	hasAllPriv := hasPriv("ALL PRIVILEGES")
+	if !hasAllPriv {
+		if !hasPriv("PROCESS") {
+			return false
+		}
+		if !hasPriv("SHOW DATABASES") {
+			return false
+		}
+		if !hasPriv("CONFIG") {
+			return false
+		}
+		hasSuperPriv := hasPriv("SUPER")
+		if !hasSuperPriv {
+			if !hasPriv("SYSTEM_VARIABLES_ADMIN") {
+				return false
+			}
+			if !hasPriv("DASHBOARD_CLIENT") {
+				return false
+			}
+		}
+	}
+	if enableSEM {
+		if !hasPriv("RESTRICTED_VARIABLES_ADMIN") {
+			return false
+		}
+		if !hasPriv("RESTRICTED_TABLES_ADMIN") {
+			return false
+		}
+		if !hasPriv("RESTRICTED_TABLES_ADMIN") {
+			return false
+		}
+	}
+	return true
 }
